chore(react-kit): deprecate withRX
